+++ scanpipe/tests/test_api.py
@@ -442,44 +442,23 @@ class ScanPipeAPITest(TransactionTestCase):
             "scan_single_package", response.data["runs"][1]["pipeline_name"]
         )
 
+        # Not supported as the comma `,` is used as the separator for optional steps.
         data = {
             "name": "Multi string",
             "pipeline": "analyze_docker_image,scan_single_package",
         }
         response = self.csrf_client.post(self.project_list_url, data)
-        self.assertEqual(status.HTTP_201_CREATED, response.status_code)
-        self.assertEqual(2, len(response.data["runs"]))
-        self.assertEqual(
-            "analyze_docker_image", response.data["runs"][0]["pipeline_name"]
-        )
-        self.assertEqual(
-            "scan_single_package", response.data["runs"][1]["pipeline_name"]
-        )
-
-        data = {
-            "name": "Mix of string and list plus selected groups",
+        self.assertEqual(status.HTTP_400_BAD_REQUEST, response.status_code)
+        expected = {
             "pipeline": [
-                "analyze_docker_image",
-                "inspect_packages:StaticResolver,scan_single_package",
-            ],
+                ErrorDetail(
+                    string='"analyze_docker_image,scan_single_package" is not a valid '
+                    "choice.",
+                    code="invalid_choice",
+                )
+            ]
         }
-        response = self.csrf_client.post(self.project_list_url, data)
-        self.assertEqual(status.HTTP_201_CREATED, response.status_code)
-        self.assertEqual(
-            "analyze_docker_image", response.data["runs"][0]["pipeline_name"]
-        )
-        self.assertEqual("inspect_packages", response.data["runs"][1]["pipeline_name"])
-        self.assertEqual(
-            "scan_single_package", response.data["runs"][2]["pipeline_name"]
-        )
-        self.assertEqual(
-            ["StaticResolver"], response.data["runs"][1]["selected_groups"]
-        )
-        runs = Project.objects.get(name=data["name"]).runs.all()
-        self.assertEqual("analyze_docker_image", runs[0].pipeline_name)
-        self.assertEqual("inspect_packages", runs[1].pipeline_name)
-        self.assertEqual("scan_single_package", runs[2].pipeline_name)
-        self.assertEqual(["StaticResolver"], runs[1].selected_groups)
+        self.assertEqual(expected, response.data)
 
     def test_scanpipe_api_project_create_pipeline_old_name_compatibility(self):
         data = {
@@ -531,6 +510,31 @@ class ScanPipeAPITest(TransactionTestCase):
         run = Project.objects.get(name="Project1").runs.get()
         self.assertEqual("inspect_packages", run.pipeline_name)
         self.assertEqual(["StaticResolver"], run.selected_groups)
+
+        data = {
+            "name": "Mix of string and list plus selected groups",
+            "pipeline": [
+                "map_deploy_to_develop:Java,JavaScript",
+                "inspect_packages:StaticResolver",
+            ],
+        }
+        response = self.csrf_client.post(self.project_list_url, data)
+        self.assertEqual(status.HTTP_201_CREATED, response.status_code)
+        self.assertEqual(
+            "map_deploy_to_develop", response.data["runs"][0]["pipeline_name"]
+        )
+        self.assertEqual("inspect_packages", response.data["runs"][1]["pipeline_name"])
+        self.assertEqual(
+            ["Java", "JavaScript"], response.data["runs"][0]["selected_groups"]
+        )
+        self.assertEqual(
+            ["StaticResolver"], response.data["runs"][1]["selected_groups"]
+        )
+        runs = Project.objects.get(name=data["name"]).runs.all()
+        self.assertEqual("map_deploy_to_develop", runs[0].pipeline_name)
+        self.assertEqual("inspect_packages", runs[1].pipeline_name)
+        self.assertEqual(["Java", "JavaScript"], runs[0].selected_groups)
+        self.assertEqual(["StaticResolver"], runs[1].selected_groups)
 
     def test_scanpipe_api_project_create_webhooks(self):
         data = {
@@ -685,6 +689,37 @@ class ScanPipeAPITest(TransactionTestCase):
         response = self.csrf_client.get(url)
         self.assertEqual("error", response.data["results"][0]["compliance_alert"])
 
+    def test_scanpipe_api_project_action_resources_filterset(self):
+        make_resource_file(
+            self.project1,
+            path="path/",
+        )
+        url = reverse("project-resources", args=[self.project1.uuid])
+        response = self.csrf_client.get(url)
+        self.assertEqual(2, response.data["count"])
+
+        response = self.csrf_client.get(url + "?path=path/")
+        self.assertEqual(1, response.data["count"])
+        package = response.data["results"][0]
+        self.assertEqual("path/", package["path"])
+
+        response = self.csrf_client.get(url + "?path=unknown")
+        self.assertEqual(0, response.data["count"])
+
+        response = self.csrf_client.get(url + "?compliance_alert=a")
+        self.assertEqual(400, response.status_code)
+        expected = {
+            "compliance_alert": [
+                "Select a valid choice. a is not one of the available choices."
+            ]
+        }
+        self.assertEqual(expected, response.data["errors"])
+
+        # Using a field name available on the Project model to make sure the
+        # ProjectFilterSet is bypassed.
+        response = self.csrf_client.get(url + "?slug=aaa")
+        self.assertEqual(2, response.data["count"])
+
     def test_scanpipe_api_project_action_packages(self):
         url = reverse("project-packages", args=[self.project1.uuid])
         response = self.csrf_client.get(url)
@@ -696,6 +731,24 @@ class ScanPipeAPITest(TransactionTestCase):
         package = response.data["results"][0]
         self.assertEqual("pkg:deb/debian/adduser@3.118?arch=all", package["purl"])
         self.assertEqual("adduser", package["name"])
+
+    def test_scanpipe_api_project_action_packages_filterset(self):
+        make_package(self.project1, package_url="pkg:generic/name@1.0")
+        url = reverse("project-packages", args=[self.project1.uuid])
+        response = self.csrf_client.get(url)
+        self.assertEqual(2, response.data["count"])
+
+        response = self.csrf_client.get(url + "?version=1.0")
+        self.assertEqual(1, response.data["count"])
+        package = response.data["results"][0]
+        self.assertEqual("pkg:generic/name@1.0", package["purl"])
+
+        response = self.csrf_client.get(url + "?version=2.0")
+        self.assertEqual(0, response.data["count"])
+
+        response = self.csrf_client.get(url + "?size=a")
+        self.assertEqual(400, response.status_code)
+        self.assertEqual({"size": ["Enter a number."]}, response.data["errors"])
 
     def test_scanpipe_api_project_action_dependencies(self):
         url = reverse("project-dependencies", args=[self.project1.uuid])
@@ -730,6 +783,15 @@ class ScanPipeAPITest(TransactionTestCase):
             "from_resource": "daglib-0.3.2.tar.gz-extract/daglib-0.3.2/PKG-INFO",
         }
         self.assertEqual(expected, relation)
+
+    def test_scanpipe_api_project_action_relations_filterset(self):
+        url = reverse("project-relations", args=[self.project1.uuid])
+        response = self.csrf_client.get(url + "?map_type=about_file")
+        self.assertEqual(0, response.data["count"])
+
+        map_type = self.codebase_relation1.map_type
+        response = self.csrf_client.get(url + f"?map_type={map_type}")
+        self.assertEqual(1, response.data["count"])
 
     def test_scanpipe_api_project_action_messages(self):
         url = reverse("project-messages", args=[self.project1.uuid])